+++ R/make_pairs.R
@@ -1,25 +1,116 @@
 #' @name make_pairs
 #' @aliases make_pairs
-#' @title Record-pair combinations.
+#' @title Combinations and permutations of record-sets.
 #'
-#' @description
-#' Create record-pair combinations of a vector's elements.
-#'
-#' @param x \code{[atomic]}.
-#' @param strata Subsets of \code{x}. A blocking attribute limiting the combinations created.
-#' @param repeats_allowed \code{[logical]} If \code{TRUE}, repetitions are included.
-#' @param permutations_allowed \code{[logical]} If \code{TRUE}, permutations are included.
+#' @param x \code{[atomic]}. Vector.
+#' @param n \code{[integer]}. Size of Vector.
+#' @param r \code{[integer]}. Number of elements in a set.
+#' @param strata Subsets of \code{x}. Blocking attribute. Limits the creation of combinations or permutations to those from the same \code{strata}.
+#' @param repeats_allowed \code{[logical]}. If \code{TRUE}, repeat values are included in each set.
+#' @param permutations_allowed \code{[logical]}. If \code{TRUE}, permutations of the same set are included.
 #' @param ... Arguments passed to \code{\link{make_pairs}}.
-#' @param data_source \code{[character]}. Data source identifier. Limits to record-pairs to those from different sources.
-#' @return A \code{list} of indexes and values of record-pair combinations.
+#' @param data_source \code{[character]}. Data source identifier. Limits the creation of combinations or permutations to those from a different \code{data_source}
+#' @return A \code{list} of a vector's elements and corresponding indexes.
+#'
+#' @details
+#' \code{\bold{set()}} - Create \code{r}-set combinations or permutations of \code{n} observations.
+#'
+#' \code{\bold{make_set()}} - Create \code{r}-set combinations or permutations of vector \code{x}.
+#'
+#' \code{\bold{make_pairs()}} - Create \code{2}-set combinations or permutations of vector \code{x}.
+#'
+#' \code{\bold{make_pairs_wf_source()}} - Create \code{2}-set combinations or permutations of vector \code{x} that are from different sources (\code{data_source}).
 #'
 #' @seealso \code{\link{eval_sub_criteria}}
 #'
 #' @examples
+#'
+#' sets(4, 2)
+#' sets(4, 2, repeats_allowed = FALSE, permutations_allowed = FALSE)
+#' make_sets(month.abb[1:4], 2)
+#' make_sets(month.abb[1:4], 3)
+#'
 #' make_pairs(month.abb[1:4])
 #' make_pairs(month.abb[1:4], strata = c(1, 1, 2, 2))
+#' make_pairs_wf_source(month.abb[1:4], data_source = c(1, 1, 2, 2))
 #'
 #' @export
+#' @rdname make_pairs
+sets <- function(n, r, permutations_allowed = TRUE, repeats_allowed = TRUE){
+  p <- n ^ r
+  pairs <- sapply(seq_len(r), function(i){
+    rev_i <- (r + 1) - i
+    V3 <- rep(seq_len(n), rep(p/(n ^ i), n))
+    V3 <- unlist(rep(list(V3), p/(n ^ rev_i)))
+    V3
+  })
+  if(isFALSE(permutations_allowed)){
+    ncols <- seq_len(ncol(pairs))[-1]
+    for (i in ncols) {
+      pairs <- pairs[!(pairs[,i] < pairs[,i-1]),]
+    }
+  }
+  if(isFALSE(repeats_allowed)){
+    ncols <- make_pairs(seq_len(ncol(pairs)),
+                        permutations_allowed = FALSE,
+                        repeats_allowed = FALSE)
+    for (i in seq_len(length(ncols$x_pos))) {
+      x_vr <- ncols$x_pos[i]
+      y_vr <- ncols$y_pos[i]
+      pairs <- pairs[which(pairs[,x_vr] != pairs[,y_vr]),]
+    }
+    rm(ncols)
+  }
+  return(pairs)
+}
+
+#' @export
+#' @rdname make_pairs
+make_sets <- function(x, r, strata = NULL, permutations_allowed = TRUE, repeats_allowed = TRUE){
+  if(!is.null(strata)){
+    strata <- match(strata, strata)
+    pos <- seq_len(length(x))
+    s_ord <- order(strata)
+    strata <- strata[s_ord]
+    pos <- pos[s_ord]
+    rl <- rle(strata)
+    n <- max(rl$lengths)
+  }else{
+    n <- length(x)
+  }
+  repo <- sets(n = n,
+               r = r,
+               permutations_allowed = permutations_allowed,
+               repeats_allowed = repeats_allowed)
+  if(!is.null(strata)){
+    sn <- seq_len(length(x))
+    rl$start_sn <- sn[!duplicated(strata)]
+    repo <- lapply(rl$values, function(x){
+      exc_indx <- which(repo > rl$lengths[rl$values == x])
+      exc_indx <- exc_indx %% nrow(repo)
+      exc_indx[exc_indx == 0] <- nrow(repo)
+      x.repo <- repo[!seq_len(nrow(repo)) %in% exc_indx,]
+      x.repo <- x.repo  + (rl$start_sn[rl$values == x] - 1)
+      x.repo
+    })
+    repo <- do.call("rbind", repo)
+    repo <- pos[repo]
+    repo <- matrix(repo, ncol = r)
+  }
+
+  repo <- split(repo,
+                rep(seq_len(r), rep(nrow(repo), r)))
+  repo <- c(repo,
+            lapply(repo, function(i) x[i]))
+  names(repo) <- paste0("x",
+                        rep(1:r, 2),
+                        rep(c("_pos", "_val"), rep(r, 2)))
+  rm(list = ls()[ls() != "repo"])
+  return(repo)
+}
+
+#' @export
+#' @rdname make_pairs
 make_pairs <- function(x, strata = NULL, repeats_allowed = TRUE, permutations_allowed = FALSE){
   # Validations
   errs <- err_make_pairs_1(x = x, strata = strata,
@@ -76,12 +167,12 @@ make_pairs <- function(x, strata = NULL, repeats_allowed = TRUE, permutations_al
 
   repo$x_val <- x[repo$x_pos]
   repo$y_val <- x[repo$y_pos]
-
-  if(!is.null(strata)){
-    repo$x_strata <- strata[repo$x_pos]
-    repo$y_strata <- strata[repo$y_pos]
-  }
   repo$ref <- repo$reps <- repo$y_max_pos <- NULL
+
+  if(isFALSE(repeats_allowed)){
+    lgk <- which(repo$x_pos != repo$y_pos)
+    repo <- lapply(repo, function(x) x[lgk])
+  }
 
   if(isTRUE(permutations_allowed)){
     lgk <- which(repo$x_pos != repo$y_pos)
@@ -89,15 +180,6 @@ make_pairs <- function(x, strata = NULL, repeats_allowed = TRUE, permutations_al
                  y_pos = c(repo$y_pos[lgk], repo$x_pos),
                  x_val = c(repo$x_val[lgk], repo$y_val),
                  y_val = c(repo$y_val[lgk], repo$x_val))
-    if(!is.null(strata)){
-      repo$x_strata <- c(repo$x_strata[lgk], repo$y_strata)
-      repo$y_strata <- c(repo$y_strata[lgk], repo$x_strata)
-    }
-  }
-
-  if(isFALSE(repeats_allowed)){
-    lgk <- which(repo$x_pos != repo$y_pos)
-    repo <- lapply(repo, function(x) x[lgk])
   }
 
   s_ord <- order(repo$y_pos, repo$x_pos)
@@ -172,3 +254,6 @@ make_pairs_wf_source <- function(..., data_source = NULL){
 
   return(repo)
 }
+
+
+